Rename TopologicalInventory to TopologicalInventory::Service
