+++ app/services/api/v1x0/catalog/provider_control_parameters.rb
@@ -12,7 +12,7 @@ module Api
 
         def process
           source_ref = PortfolioItem.find(@portfolio_item_id).service_offering_source_ref
-          TopologicalInventory.call do |api_instance|
+          TopologicalInventory::Service.call do |api_instance|
             # TODO: Temporay till we get this call in the topology service
             projects = api_instance.list_source_container_projects(source_ref).data
             update_project_list(projects)
