+++ app/services/api/v1x0/catalog/service_offering.rb
@@ -11,7 +11,7 @@ module Api
         end
 
         def process
-          service_offering = TopologicalInventory.call do |api|
+          service_offering = TopologicalInventory::Service.call do |api|
             api.show_service_offering(@service_offering_ref)
           end
