+++ app/services/api/v1x0/service_offering/add_to_portfolio_item.rb
@@ -13,7 +13,7 @@ module Api
 
         def process
           # Get the api response object from Topology
-          TopologicalInventory.call do |api_instance|
+          TopologicalInventory::Service.call do |api_instance|
             @service_offering = api_instance.show_service_offering(@params[:service_offering_ref])
           end
 
@@ -54,7 +54,7 @@ module Api
         end
 
         def create_icon(icon_id)
-          service_offering_icon = TopologicalInventory.call { |topo| topo.show_service_offering_icon(icon_id) }
+          service_offering_icon = TopologicalInventory::Service.call { |topo| topo.show_service_offering_icon(icon_id) }
           return if service_offering_icon.data.nil?
 
           file = Tempfile.new('service_offering').tap do |fh|
