+++ app/services/api/v1x0/tags/topology/remote_inventory.rb
@@ -33,7 +33,7 @@ module Api
 
           def all_tag_collections
             @task.context[:applied_inventories].collect do |inventory_id|
-              TopologicalInventory.call do |api|
+              TopologicalInventory::Service.call do |api|
                 api.list_service_inventory_tags(inventory_id).data
               end
             end
